+++ api/lib/ZwaveClient.ts
@@ -16,19 +16,19 @@ import {
 	ZWaveDataRate,
 	ZWaveErrorCodes,
 	Protocols,
-	createDefaultTransportFormat,
 	FirmwareFileFormat,
 	tryUnzipFirmwareFile,
-	extractFirmwareAsync,
+	extractFirmware,
 } from '@zwave-js/core'
+import { createDefaultTransportFormat } from '@zwave-js/core/bindings/log/node'
 import { JSONTransport } from '@zwave-js/log-transport-json'
-import { isDocker } from '@zwave-js/shared'
+import { isDocker } from './utils'
 import {
 	AssociationAddress,
 	AssociationGroup,
-	ControllerFirmwareUpdateProgress,
-	ControllerFirmwareUpdateResult,
-	ControllerFirmwareUpdateStatus,
+	OTWFirmwareUpdateProgress,
+	OTWFirmwareUpdateResult,
+	OTWFirmwareUpdateStatus,
 	ControllerStatistics,
 	ControllerStatus,
 	DataRate,
@@ -108,6 +108,7 @@ import {
 	ProvisioningEntryStatus,
 	AssociationCheckResult,
 	LinkReliabilityCheckResult,
+	DriverMode,
 } from 'zwave-js'
 import { getEnumMemberName, parseQRCodeString } from 'zwave-js/Utils'
 import { configDbDir, logsDir, nvmBackupsDir, storeDir } from '../config/app'
@@ -127,8 +128,10 @@ import { readFile } from 'fs/promises'
 import backupManager, { NVM_BACKUP_PREFIX } from './BackupManager'
 import { socketEvents } from './SocketEvents'
 import { isUint8Array } from 'util/types'
+import { PkgFsBindings } from './PkgFsBindings'
+import { join } from 'path'
 
-export const deviceConfigPriorityDir = storeDir + '/config'
+export const deviceConfigPriorityDir = join(storeDir, 'config')
 
 export const configManager = new ConfigManager({
 	deviceConfigPriorityDir,
@@ -175,6 +178,7 @@ export const allowedApis = validateMethods([
 	'pollValue',
 	'setPowerlevel',
 	'setRFRegion',
+	'setMaxLRPowerLevel',
 	'updateControllerNodeProps',
 	'startInclusion',
 	'startExclusion',
@@ -465,7 +469,7 @@ export interface BackgroundRSSIPoint {
 
 export interface FwFile {
 	name: string
-	data: Buffer | Uint8Array
+	data: Uint8Array
 	target?: number
 }
 
@@ -531,6 +535,7 @@ export type ZUINode = {
 	isControllerNode?: boolean
 	powerlevel?: number
 	measured0dBm?: number
+	maxLongRangePowerlevel?: number
 	RFRegion?: RFRegion
 	rfRegions?: { text: string; value: number }[]
 	isFrequentListening?: FLiRS
@@ -621,6 +626,7 @@ export type ZwaveConfig = {
 	disableControllerRecovery?: boolean
 	rf?: {
 		region?: RFRegion
+		maxLongRangePowerlevel?: number
 		txPower?: {
 			powerlevel: number
 			measured0dBm: number
@@ -2137,7 +2143,7 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 
 		// extend options with hidden `options`
 		const zwaveOptions: PartialZWaveOptions = {
-			allowBootloaderOnly: this.cfg.allowBootloaderOnly || false,
+			bootloaderMode: this.cfg.allowBootloaderOnly ? 'allow' : 'recover',
 			storage: {
 				cacheDir: storeDir,
 				deviceConfigPriorityDir:
@@ -2187,12 +2193,16 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 		}
 
 		if (this.cfg.rf) {
-			const { region, txPower } = this.cfg.rf
+			const { region, txPower, maxLongRangePowerlevel } = this.cfg.rf
 
 			zwaveOptions.rf = {}
 
 			if (typeof region === 'number') {
 				zwaveOptions.rf.region = region
+			}
+
+			if (typeof maxLongRangePowerlevel === 'number') {
+				zwaveOptions.rf.maxLongRangePowerlevel = maxLongRangePowerlevel
 			}
 
 			if (
@@ -2202,6 +2212,14 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 			) {
 				zwaveOptions.rf.txPower = txPower
 			}
+		}
+
+		// @ts-expect-error this is defined when running in a pkg bundle
+		if (process.pkg) {
+			// Ensure Z-Wave JS is looking for the configuration files in the right place
+			// when running inside a pkg bundle
+			zwaveOptions.host ??= {}
+			zwaveOptions.host.fs = new PkgFsBindings()
 		}
 
 		// ensure deviceConfigPriorityDir exists to prevent warnings #2374
@@ -2274,6 +2292,14 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 			this._driver.on(
 				'bootloader ready',
 				this._onBootLoaderReady.bind(this),
+			)
+			this._driver.on(
+				'firmware update progress',
+				this._onOTWFirmwareUpdateProgress.bind(this),
+			)
+			this._driver.on(
+				'firmware update finished',
+				this._onOTWFirmwareUpdateFinished.bind(this),
 			)
 
 			logger.info(`Connecting to ${this.cfg.port}`)
@@ -2981,7 +3007,7 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 			if (strategy === InclusionStrategy.Security_S2) {
 				let inclusionOptions: ReplaceNodeOptions
 				if (options?.qrString) {
-					const parsedQr = parseQRCodeString(options.qrString)
+					const parsedQr = await parseQRCodeString(options.qrString)
 
 					if (parsedQr) {
 						// when replacing a failed node you cannot use smart start so always use qrcode for provisioning
@@ -3145,6 +3171,21 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 		throw new DriverNotReadyError()
 	}
 
+	async setMaxLRPowerLevel(powerlevel: number): Promise<boolean> {
+		if (this.driverReady) {
+			const result =
+				await this._driver.controller.setMaxLongRangePowerlevel(
+					powerlevel,
+				)
+			await this.updateControllerNodeProps(null, [
+				'maxLongRangePowerlevel',
+			])
+			return result
+		}
+
+		throw new DriverNotReadyError()
+	}
+
 	/**
 	 * Start inclusion
 	 */
@@ -3206,7 +3247,9 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 						break
 					case InclusionStrategy.Security_S2:
 						if (options?.qrString) {
-							const parsedQr = parseQRCodeString(options.qrString)
+							const parsedQr = await parseQRCodeString(
+								options.qrString,
+							)
 							if (!parsedQr) {
 								throw Error(`Invalid QR code string`)
 							}
@@ -3216,7 +3259,7 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 							} else if (
 								parsedQr.version === QRCodeVersion.SmartStart
 							) {
-								this.provisionSmartStartNode(parsedQr)
+								await this.provisionSmartStartNode(parsedQr)
 								return true
 							} else {
 								throw Error(`Invalid QR code version`)
@@ -3882,9 +3925,7 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 	/**
 	 * Used to trigger an update of controller FW
 	 */
-	async firmwareUpdateOTW(
-		file: FwFile,
-	): Promise<ControllerFirmwareUpdateResult> {
+	async firmwareUpdateOTW(file: FwFile): Promise<OTWFirmwareUpdateResult> {
 		try {
 			if (backupManager.backupOnEvent) {
 				this.nvmEvent = 'before_controller_fw_update_otw'
@@ -3894,15 +3935,13 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 
 			try {
 				const format = guessFirmwareFileFormat(file.name, file.data)
-				firmware = await extractFirmwareAsync(file.data, format)
+				firmware = await extractFirmware(file.data, format)
 			} catch (err) {
 				throw Error(
 					`Unable to extract firmware from file '${file.name}'`,
 				)
 			}
-			const result = await this.driver.controller.firmwareUpdateOTW(
-				firmware.data,
-			)
+			const result = await this.driver.firmwareUpdateOTW(firmware.data)
 			return result
 		} catch (e) {
 			throw Error(`Error while updating firmware: ${e.message}`)
@@ -3939,7 +3978,7 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 
 			for (const f of files) {
 				let { data, name } = f
-				if (data instanceof Buffer) {
+				if (isUint8Array(data)) {
 					try {
 						let format: FirmwareFileFormat
 						if (name.endsWith('.zip')) {
@@ -3957,10 +3996,7 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 							format = guessFirmwareFileFormat(name, data)
 						}
 
-						const firmware = await extractFirmwareAsync(
-							data,
-							format,
-						)
+						const firmware = await extractFirmware(data, format)
 						if (f.target !== undefined) {
 							firmware.firmwareTarget = f.target
 						}
@@ -4143,7 +4179,7 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 
 		logger.log('info', 'Calling api %s with args: %o', apiName, args)
 
-		if (this.driverReady || this.driver?.isInBootloader()) {
+		if (this.driverReady || this.driver?.mode === DriverMode.Bootloader) {
 			try {
 				const allowed =
 					typeof this[apiName] === 'function' &&
@@ -4419,14 +4455,6 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 						this._onControllerStatisticsUpdated.bind(this),
 					)
 					.on(
-						'firmware update progress',
-						this._onControllerFirmwareUpdateProgress.bind(this),
-					)
-					.on(
-						'firmware update finished',
-						this._onControllerFirmwareUpdateFinished.bind(this),
-					)
-					.on(
 						'status changed',
 						this._onControllerStatusChanged.bind(this),
 					)
@@ -4503,75 +4531,47 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 		}
 	}
 
-	private _onControllerFirmwareUpdateProgress(
-		progress: ControllerFirmwareUpdateProgress,
-	) {
-		const nodeId = this.driver.controller.ownNodeId
-		const node = this.nodes.get(nodeId)
-		if (node) {
-			node.firmwareUpdate = {
-				sentFragments: progress.sentFragments,
-				totalFragments: progress.totalFragments,
-				progress: progress.progress,
-				currentFile: node.firmwareUpdate?.currentFile ?? 1,
-				totalFiles: node.firmwareUpdate?.currentFile ?? 1,
-			}
-
-			// send at most 4msg per second
-			this.throttle(
-				this._onControllerFirmwareUpdateProgress.name,
-				this.emitNodeUpdate.bind(this, node, {
-					firmwareUpdate: node.firmwareUpdate,
-				} as utils.DeepPartial<ZUINode>),
-				250,
-			)
-		}
+	private _onOTWFirmwareUpdateProgress(progress: OTWFirmwareUpdateProgress) {
+		this.throttle(
+			this._onOTWFirmwareUpdateProgress.name,
+			this.sendToSocket.bind(this, socketEvents.otwFirmwareUpdate, {
+				progress,
+			}),
+			250,
+		)
 
 		this.emit(
 			'event',
-			EventSource.CONTROLLER,
+			EventSource.DRIVER,
 			'controller firmware update progress',
-			this.zwaveNodeToJSON(this.driver.controller.nodes.get(nodeId)),
 			progress,
 		)
 	}
 
-	private _onControllerFirmwareUpdateFinished(
-		result: ControllerFirmwareUpdateResult,
-	) {
-		const nodeId = this.driver.controller.ownNodeId
-		const node = this.nodes.get(nodeId)
-		const zwaveNode = this.driver.controller.nodes.get(nodeId)
-
-		if (node) {
-			node.firmwareUpdate = undefined
-
-			this.emitNodeUpdate(node, {
-				firmwareUpdate: false,
-				firmwareUpdateResult: {
-					success: result.success,
-					status: getEnumMemberName(
-						ControllerFirmwareUpdateStatus,
-						result.status,
-					),
-				},
-			} as any)
-		}
+	private _onOTWFirmwareUpdateFinished(result: OTWFirmwareUpdateResult) {
+		this.sendToSocket(socketEvents.otwFirmwareUpdate, {
+			result: {
+				success: result.success,
+				status: getEnumMemberName(
+					OTWFirmwareUpdateStatus,
+					result.status,
+				),
+			},
+		})
 
 		logger.info(
-			`Controller ${zwaveNode.id} firmware update OTW finished ${
+			`Controller firmware update OTW finished ${
 				result.success ? 'successfully' : 'with error'
 			}.\n   Status: ${getEnumMemberName(
-				ControllerFirmwareUpdateStatus,
+				OTWFirmwareUpdateStatus,
 				result.status,
 			)}. Result: ${JSON.stringify(result)}.`,
 		)
 
 		this.emit(
 			'event',
-			EventSource.CONTROLLER,
+			EventSource.DRIVER,
 			'controller firmware update finished',
-			this.zwaveNodeToJSON(zwaveNode),
 			result,
 		)
 	}
@@ -5007,7 +5007,7 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 		this._updateControllerStatus(`Backup NVM progress: ${progress}%`)
 	}
 
-	async restoreNVM(data: Buffer, useRaw = false) {
+	async restoreNVM(data: Uint8Array, useRaw = false) {
 		if (!this.driverReady) {
 			throw new DriverNotReadyError()
 		}
@@ -5093,12 +5093,12 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 		this.driver.controller.unprovisionSmartStartNode(dskOrNodeId)
 	}
 
-	parseQRCodeString(qrString: string): {
+	async parseQRCodeString(qrString: string): Promise<{
 		parsed?: QRProvisioningInformation
 		nodeId?: number
 		exists: boolean
-	} {
-		const parsed = parseQRCodeString(qrString)
+	}> {
+		const parsed = await parseQRCodeString(qrString)
 		let node: ZWaveNode | undefined
 		let exists = false
 
@@ -5117,14 +5117,14 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 		}
 	}
 
-	provisionSmartStartNode(entry: PlannedProvisioningEntry | string) {
+	async provisionSmartStartNode(entry: PlannedProvisioningEntry | string) {
 		if (!this.driverReady) {
 			throw new DriverNotReadyError()
 		}
 
 		if (typeof entry === 'string') {
 			// it's a qrcode
-			entry = parseQRCodeString(entry)
+			entry = await parseQRCodeString(entry)
 		}
 
 		if (!entry.dsk) {
@@ -5719,10 +5719,9 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 		} else if (ccId === CommandClasses['Entry Control']) {
 			valueId.property = args.eventType.toString()
 			valueId.propertyKey = args.dataType
-			data =
-				args.eventData instanceof Buffer
-					? utils.buffer2hex(args.eventData)
-					: args.eventData
+			data = isUint8Array(args.eventData)
+				? utils.buffer2hex(args.eventData)
+				: args.eventData
 		} else if (ccId === CommandClasses['Multilevel Switch']) {
 			valueId.property = getEnumMemberName(
 				MultilevelSwitchCommand,
@@ -6189,7 +6188,11 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 
 	async updateControllerNodeProps(
 		node?: ZUINode,
-		props: Array<'powerlevel' | 'RFRegion'> = ['powerlevel', 'RFRegion'],
+		props: Array<'powerlevel' | 'RFRegion' | 'maxLongRangePowerlevel'> = [
+			'powerlevel',
+			'RFRegion',
+			'maxLongRangePowerlevel',
+		],
 	) {
 		node = node || this.nodes.get(this._driver.controller.ownNodeId)
 		if (props.includes('powerlevel')) {
@@ -6228,11 +6231,26 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 			}
 		}
 
+		if (props.includes('maxLongRangePowerlevel')) {
+			if (
+				this._driver.controller.isSerialAPISetupCommandSupported(
+					SerialAPISetupCommand.GetLongRangeMaximumTxPower,
+				)
+			) {
+				const limit =
+					await this._driver.controller.getMaxLongRangePowerlevel()
+				node.maxLongRangePowerlevel = limit
+			} else {
+				logger.info('LR powerlevel is not supported by controller')
+			}
+		}
+
 		this.emitNodeUpdate(node, {
 			powerlevel: node.powerlevel,
 			measured0dBm: node.measured0dBm,
 			RFRegion: node.RFRegion,
 			supportsLongRange: node.supportsLongRange,
+			maxLongRangePowerlevel: node.maxLongRangePowerlevel,
 		})
 	}
 
@@ -6754,8 +6772,8 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 				let api: 'firmwareUpdateOTW' | 'firmwareUpdateOTA'
 				if (this.nodes.get(nodeId).isControllerNode) {
 					api = 'firmwareUpdateOTW'
-					this._onControllerFirmwareUpdateFinished({
-						status: ControllerFirmwareUpdateStatus.OK,
+					this._onOTWFirmwareUpdateFinished({
+						status: OTWFirmwareUpdateStatus.OK,
 						success: true,
 					})
 				} else {
@@ -6799,7 +6817,7 @@ class ZwaveClient extends TypedEventEmitter<ZwaveClientEventCallbacks> {
 					.catch(() => {
 						//noop
 					})
-				this._onControllerFirmwareUpdateProgress({
+				this._onOTWFirmwareUpdateProgress({
 					sentFragments: progress.sentFragments,
 					totalFragments: progress.totalFragments,
 					progress: progress.progress,