+++ src/lib/items.js
@@ -1,5 +1,9 @@
-import { RFRegion, Protocols } from 'zwave-js/safe'
-import { ZnifferRegion } from '@zwave-js/core/safe'
+import {
+	ZnifferLRChannelConfig,
+	ZnifferRegion,
+	RFRegion,
+	Protocols,
+} from '@zwave-js/core'
 
 export const rfRegions = Object.keys(RFRegion)
 	.filter((k) => isNaN(k))
@@ -23,6 +27,13 @@ export const znifferRegions = Object.keys(ZnifferRegion)
 	}))
 	.sort((a, b) => a.text.localeCompare(b.text))
 
+export const znifferLRChannelConfigs = Object.keys(ZnifferLRChannelConfig)
+	.filter((k) => isNaN(k))
+	.map((key) => ({
+		text: key,
+		value: ZnifferLRChannelConfig[key],
+	}))
+
 export const protocolsItems = [
 	{
 		text: 'Z-Wave',
@@ -31,5 +42,16 @@ export const protocolsItems = [
 	{
 		text: 'Z-Wave Long Range',
 		value: Protocols.ZWaveLongRange,
+	},
+]
+
+export const maxLRPowerLevels = [
+	{
+		text: '+14 dBm',
+		value: 14,
+	},
+	{
+		text: '+20 dBm',
+		value: 20,
 	},
 ]