+++ src/lib/utils.js
@@ -2,15 +2,12 @@ import {
 	isValidDSK,
 	Protocols,
 	znifferProtocolDataRateToString,
-} from '@zwave-js/core/safe'
-
-import {
 	isRssiError,
 	rssiToString,
-	getEnumMemberName,
-	ZWaveFrameType,
-	LongRangeFrameType,
-} from 'zwave-js/safe'
+} from '@zwave-js/core'
+
+import { getEnumMemberName } from '@zwave-js/shared'
+import { ZWaveFrameType, LongRangeFrameType } from 'zwave-js'
 import { znifferRegions } from './items'
 import { mdiZWave } from '@mdi/js'